+++ backend/src/controllers/auth.controllers.js
@@ -5,10 +5,10 @@ import generateToken from '../helpers/generateToken.js';
 export async function register(req, res) {
     try {
         const { username, password } = req.body;
-        if (!username || !password) return res.status(400).json({ error: 'Username and password are required' });
+        if (!username || !password) return res.status(400).json({ message: 'Username and password are required' });
 
         const dbUser = await User.findOne({ username });
-        if (dbUser) return res.status(409).json({ message: 'Username already exists' });
+        if (dbUser) return res.status(409).json({ message: 'User already exists' });
 
         const _id = new mongoose.Types.ObjectId();
         const lastUser = await User.findOne().sort({ _id: -1 }).exec();
@@ -27,13 +27,15 @@ export async function register(req, res) {
 export async function login(req, res) {
     try {
         const { username, password } = req.body;
-        if (!username || !password) return res.status(400).json({ error: 'Username and password are required' });
+        if (!username || !password) return res.status(400).json({ message: 'Username and password are required' });
 
         const dbUser = await User.findOne({ username });
-        if (!dbUser) return res.status(401).json({ error: 'Invalid credentials' });
-        
-        const isPasswordCorrect = await dbUser.checkPassword(password);
-        if (!isPasswordCorrect) return res.status(401).json({ error: 'Invalid credentials' });
+        if (!dbUser) return res.status(401).json({ message: 'Invalid credentials' });
+
+        if (dbUser.password) {
+            const isPasswordCorrect = await dbUser.checkPassword(password);
+            if (!isPasswordCorrect) return res.status(401).json({ message: 'Invalid credentials' });
+        } else return res.status(403).json({ message: 'Invalid permissions' });
 
         const _id = dbUser._id;
         const token = generateToken(_id);