+++ backend/src/middleware/admin.middleware.js
@@ -1,0 +1,10 @@
+export default async function (req, res, next) {
+    try {
+        const user = req.user;
+        if (!user.is_admin) throw new Error('Invalid permissions');
+
+        next();
+    } catch (err) {
+        return res.status(403).json({ message: err.message }); 
+    }
+};