+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 #
 **OO-0001**<br>
-Add register (backend+frontend)<br>
-Add login (backend+frontend)<br>
+* Add register (backend+frontend)<br>
+* Add login (backend+frontend)<br>
+
+**OO-0002**<br>
+* Add is_admin flag to user schema (backend)<br>
+* Add is_deleted flag to user schema (backend)<br>
+* Add users CRUD (backend)<br>
+* Secure routes (is logged / is admin) (backend)<br>
+* Add user to req in secured routes (backend)<br>
 #